Fixed infinite loop of asking for a direction when no directions are possible
Fixed the infinite loop bug. Now the program checks if any direction is possible right after a player chooses a grid space, and if none are possible throws and exception that allows the player to choose a new space. Also, changed one typo and changed the text between turns when placing ships so a player knows if they have another ship to place.